requirements.txt
appnope==0.1.3
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
evernote3==1.25.14
gcloud==0.18.3
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1
httplib2==0.20.4
idna==3.3
ipdb==0.13.9
ipython==7.33.0
jedi==0.18.1
lxml==4.8.0
matplotlib-inline==0.1.3
oauth2==1.9.0.post1
oauth2client==4.1.3
oauthlib==3.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
protobuf==3.20.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.9
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
toml==0.10.2
traitlets==5.2.0
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
